refactor(orc8r): Device service security changes
